+++ mephisto/client/cli.py
@@ -1,5 +1,5 @@
 import click
-from mephisto.cli.server import app
+from mephisto.client.server import app
 from click_default_group import DefaultGroup
 
 @click.group(cls=DefaultGroup, default='web', default_if_no_args=True)
