+++ mephisto/client/server.py
@@ -1,7 +1,7 @@
 from flask import Flask, send_file, jsonify
 import os
-from mephisto.cli.config import Config
-from mephisto.cli.api import api
+from mephisto.client.config import Config
+from mephisto.client.api import api
 
 app = Flask(__name__, static_url_path="/static", static_folder="../webapp/build/static")
 app.config.from_object(Config)
