+++ pyproject.toml
@@ -14,7 +14,7 @@ click_default_group = "^1.2"
 pytest = "^3.0"
 
 [tool.poetry.scripts]
-'mephisto' = 'mephisto.cli.client:cli'
+'mephisto' = 'mephisto.client.cli:cli'
 
 [build-system]
 requires = ["poetry>=0.12"]
